Plotting
I didn't figure out how to use the ReadNodeFile without taking in extra stuff I didn't need for the plots, so I created a new function to just take the node values. I think we can work on merging bit of the plot with the triangle class, but that's something for later. For now this bit works.

Sorry for the delay. :)